% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgrid.R
\name{getPrioRaster}
\alias{getPrioRaster}
\title{Get PRIO GRID Data as a Raster}
\usage{
getPrioRaster(names, years = c())
}
\arguments{
\item{names}{The names of the variables to be transformed into a raster. A warning will be issues for variables
that are unavailable.}

\item{years}{An integer vector indicating the years for which the yearly variables should be returned. Ignored for
static variables.}
}
\value{
A list containing (1) a VeloxRaster object and (2) a data frame with meta data about the PRIO GRID variables.
}
\description{
Returns the requested PRIO GRID variables as a global raster at 0.5 x 0.5 decimal degree resolution.
}
\details{
Note that only variables contained in the PRIO GRID csv files linked by the \code{setDataPath} function will
be available. Download the PRIO GRID csv files at \url{http://http://grid.prio.org/#/download}.
}
\examples{

\dontrun{
prio.ls <- getPrioRaster(names = 'nlights_calib_mean', years = 2005:2010)
vx <- prio.ls$raster
meta.df <- prio.ls$meta
}

}

