% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgrid.R
\name{getVarNames}
\alias{getVarNames}
\title{Get Available PRIO GRID Variables}
\usage{
getVarNames(type = c("static", "yearly"), details = FALSE)
}
\arguments{
\item{type}{A character string, either \code{'static'} or \code{'yearly'}.}

\item{details}{A boolean. If \code{TRUE}, returns additional details about the available variables (sources, units, etc).}
}
\value{
A data frame.
}
\description{
\code{getVarNames} returns a data frame of all PRIO GRID variables available to the package.
Which variables are available depends on the contents of the PRIO GRID csv files linked by the
\code{setDataPath} function.
}
\examples{

\dontrun{
vars.df <- getVarNames(type = 'static', details = TRUE)
}

}

