% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgrid.R
\docType{data}
\name{priogrid.df}
\alias{priogrid.df}
\title{This is data to be included in my package}
\description{
This is data to be included in my package
}
\author{
PRIO GRID
}
\references{
\url{http://grid.prio.org}
}
\keyword{GRID}
\keyword{PRIO}
\keyword{description}
\keyword{variable}

