% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgrid.R
\name{setDataPath}
\alias{setDataPath}
\title{Set Paths to PRIO GRID csv Files}
\usage{
setDataPath(staticPath = "", yearlyPath = "")
}
\arguments{
\item{staticPath}{A character string represeting a path pointing to a static PRIO GRID csv file.
If the default value of \code{''} is passed, no changes are made.}

\item{yearlyPath}{A character string represeting a path pointing to a yearly PRIO GRID csv file.
If the default value of \code{''} is passed, no changes are made.}
}
\value{
Void.
}
\description{
\code{setDataPath} is used to (re)set the paths to the csv files containing the
PRIO GRID raw data. The path information is saved in the installation directory
of the package, so it is persistent across R sessions.
}
\details{
It is possible to set a single path if only the static or the yearly variables are required.
Also, (re)setting a single path will not overwrite the other.
}
\examples{

\dontrun{
setDataPath(staticPath = '/home/usr/data/PRIO-GRID Static Variables - 2017-12-08.csv')
}

}

